Giv MemberAdmin-objektet sin egen fil
